+++ internal/dot11build/deauth_build.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package frame80211
+package dot11build
 
 import (
 	"encoding/binary"
