+++ internal/dot11dissec/data_frame_dissec.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package frame80211
+package dot11dissec
