+++ internal/dot11dissec/dot11_dissec.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package frame80211
+package dot11dissec
 
 import "encoding/binary"
