+++ internal/pktbuild/ip_header.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package packet
+package pktbuild
 
 import (
 	"encoding/binary"
