+++ internal/pktdissec/ether_dissec.go
@@ -15,20 +15,16 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package frame80211
+package pktdissec
+
+import "net"
 
 
-func minimalRariotapHeader(buffer []byte) {
-	buffer[0]  = 0x00  // Header revision
-	buffer[1]  = 0x00  // Header pad
-	buffer[2]  = 0x0c  // Header length
-	buffer[3]  = 0x00  //
-	buffer[4]  = 0x04  // Bitmap
-	buffer[5]  = 0x80  //
-	buffer[6]  = 0x00  //
-	buffer[7]  = 0x00  //
-	buffer[8]  = 0x02  // Rate
-	buffer[9]  = 0x00  // Rate pad
-	buffer[10] = 0x18  // TX flags
-	buffer[11] = 0x00  //
+
+func (pd *PacketDissector) GetEtherSrcMAC() (net.HardwareAddr, bool) {
+    if pd.lenPkt < 12 {
+        return nil, false
+    }
+    
+	return net.HardwareAddr(pd.pkt[6:12]), true
 }
