+++ internal/pktdissec/ip_dissec.go
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org>.
  */
 
-package packet
+package pktdissec
 
 import (
 	"encoding/binary"
